// frontend/src/client/models/AnnouncementCreate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type AnnouncementCreate = {
    id?: (number | null);
    course_id: number;
    content?: (string | null);
    announcement_time?: (string | null);
    course_location?: (string | null);
    course_time?: (string | null);
    course_status?: (string | null);
};